platformize oauth page
