+++ .eslintrc
@@ -15,6 +15,7 @@
   ],
   "rules": {
     "max-len": ["error", { "code": 100 }],
+    "no-prototype-builtins": "off",
     "react/jsx-uses-react": "off",
     "react/react-in-jsx-scope": "off",
     "@typescript-eslint/no-empty-function": "off",
