+++ src/manifest.json
@@ -26,6 +26,6 @@
     "chrome_style": true,
     "page": "./pages/options.html"
   },
-  "web_accessible_resources": ["./services/pocket/oauth.html"],
+  "web_accessible_resources": ["./pages/oauth/pocket.html"],
   "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'"
 }
