+++ src/pages/oauth/pocket.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>Pile oAuth page for Pocket</title>
+</head>
+
+<body>
+  <script src="./index.tsx" type="module"></script>
+  <div id="root" service="pocket"></div>
+</body>
+
+</html>
