--- src/services/pocket/oauth.html
@@ -1,13 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-
-<head>
-  <title>Pile oAuth page for Pocket</title>
-</head>
-
-<body>
-  <script src="./oauth.tsx" type="module"></script>
-  <div id="root"></div>
-</body>
-
-</html>
